TaskRunLog: Refactor to separate package
- Crate new apis/v1alpha2 package to hold APIs internal to Results
- Move TaskRunLog to apis/v1alpha2 package
